# backend/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"MessageResponse": {
				"description": "Generic success envelope",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InternshipBody": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"description": "ACADEMIC | PROFESSIONAL"
					},
					"field": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"isRemote": {
						"type": "boolean"
					},
					"isPaid": {
						"type": "boolean"
					},
					"stipendAmount": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"durationWeeks": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"applicationDeadline": {
						"type": "string"
					},
					"slots": {
						"type": "number",
						"format": "double"
					},
					"requirements": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"responsibilities": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"skills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"type",
					"field"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_InternshipBody_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"description": "ACADEMIC | PROFESSIONAL"
					},
					"field": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"isRemote": {
						"type": "boolean"
					},
					"isPaid": {
						"type": "boolean"
					},
					"stipendAmount": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"durationWeeks": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"applicationDeadline": {
						"type": "string"
					},
					"slots": {
						"type": "number",
						"format": "double"
					},
					"requirements": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"responsibilities": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"skills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"EvaluationBody": {
				"properties": {
					"enrollmentId": {
						"type": "string"
					},
					"stage": {
						"type": "string",
						"description": "MIDTERM | FINAL"
					},
					"punctuality": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"communication": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"technicalSkills": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"teamwork": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"initiative": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"professionalism": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"strengths": {
						"type": "string"
					},
					"areasForImprovement": {
						"type": "string"
					},
					"generalComments": {
						"type": "string"
					},
					"recommendForHire": {
						"type": "boolean"
					}
				},
				"required": [
					"enrollmentId",
					"stage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SelfAssessmentBody": {
				"properties": {
					"enrollmentId": {
						"type": "string"
					},
					"stage": {
						"type": "string",
						"description": "MIDTERM | FINAL"
					},
					"skillsDeveloped": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"goalsMet": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"supervisorSupport": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"workEnvironment": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"overallExperience": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"achievements": {
						"type": "string"
					},
					"challenges": {
						"type": "string"
					},
					"lessonsLearned": {
						"type": "string"
					},
					"futureGoals": {
						"type": "string"
					}
				},
				"required": [
					"stage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateStatusBody": {
				"properties": {
					"status": {
						"type": "string",
						"description": "APPLIED | REVIEWED | INTERVIEW_SCHEDULED | ACCEPTED | REJECTED | WITHDRAWN"
					},
					"rejectionReason": {
						"type": "string"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ScheduleInterviewBody": {
				"properties": {
					"scheduledAt": {
						"type": "string",
						"description": "ISO 8601 datetime string"
					},
					"durationMinutes": {
						"type": "number",
						"format": "double"
					},
					"meetingLink": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"interviewerName": {
						"type": "string"
					}
				},
				"required": [
					"scheduledAt"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Internship Monitoring & Connecting System API",
		"version": "1.0.0",
		"description": "Full-stack platform bridging students, companies, and supervisors throughout the internship lifecycle.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/internships": {
			"get": {
				"operationId": "ListInternships",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Browse all open internship listings (public)",
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "field",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isRemote",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "isPaid",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateInternship",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Create a new internship listing (Company only)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InternshipBody"
							}
						}
					}
				}
			}
		},
		"/api/internships/{id}": {
			"get": {
				"operationId": "GetInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get a single internship by ID (public)",
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Update an internship listing (Company only)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_InternshipBody_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Delete an internship listing (Company or Admin)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/internships/{id}/bookmark": {
			"post": {
				"operationId": "ToggleBookmark",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Toggle bookmark on an internship (Student only)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/internships/bookmarks/me": {
			"get": {
				"operationId": "GetBookmarks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get current student's bookmarked internships",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/internships/company/mine": {
			"get": {
				"operationId": "GetMyInternships",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get internships posted by the authenticated company",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/evaluations": {
			"post": {
				"operationId": "SubmitEvaluation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Submit a supervisor evaluation for an enrolled student.\r\nRequires ACADEMIC_SUPERVISOR or SITE_SUPERVISOR role.",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EvaluationBody"
							}
						}
					}
				}
			}
		},
		"/api/evaluations/enrollment/{enrollmentId}": {
			"get": {
				"operationId": "GetEvaluations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all evaluations for a specific enrollment",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "enrollmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/evaluations/{id}/approve": {
			"patch": {
				"operationId": "ApproveEvaluation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Approve a submitted evaluation (Admin only)",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/evaluations/self-assessment": {
			"post": {
				"operationId": "SubmitSelfAssessment",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Submit a student self-assessment (Student only)",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SelfAssessmentBody"
							}
						}
					}
				}
			}
		},
		"/api/evaluations/self-assessment/me": {
			"get": {
				"operationId": "GetMySelfAssessments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all self-assessments for the authenticated student",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/applications/me": {
			"get": {
				"operationId": "GetMyApplications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all applications submitted by the authenticated student",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/applications/{id}/withdraw": {
			"patch": {
				"operationId": "Withdraw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Withdraw a submitted application (Student only)",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/applications/internship/{internshipId}": {
			"get": {
				"operationId": "GetApplicationsForInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all applications for a specific internship (Company only)",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "internshipId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/applications/{id}/status": {
			"patch": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Update application status — accept, reject, or move to review (Company only)",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStatusBody"
							}
						}
					}
				}
			}
		},
		"/api/applications/{id}/interview": {
			"post": {
				"operationId": "ScheduleInterview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Schedule an interview for an application (Company only)",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ScheduleInterviewBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}